+++ src/main/kotlin/klib/files/Directory.kt
@@ -1,0 +1,12 @@
+package klib.files
+
+import klib.annotations.Experimental
+
+/**
+ * Generate a random string with only Base58 characters
+ *
+ * @since 5.1.0 (Experimental)
+ * @author Thomas Obernosterer
+ */
+@OptIn(Experimental::class)
+typealias Directory = Directory2